refactor(scan-from-file): move to pipelines folder
