+++ src/cli/execution.test.ts
@@ -1,13 +1,13 @@
 import { mocked } from 'ts-jest/utils'
 import meow from 'meow'
 
-import scanFromFile from '../scanFromFile'
+import scanFromFile from '../pipelines/scanFromFile'
 import execution from './execution'
 
 const meowMocked = mocked(meow, true)
 const scanFromFileMocked = mocked(scanFromFile, true)
 
-jest.mock('../scanFromFile', jest.fn)
+jest.mock('../pipelines/scanFromFile', jest.fn)
 jest.mock('meow', jest.fn)
 
 const DEFAULT_MEOW_PROPERTIES = {
